// src/app/model/movieWithPoster.ts
export class MovieWithPoster {
    movieName: string = '';
    movieDirector: string = '';
    movieGenre: string = '';
    movieReleaseDate: string = '';
    movieLanguage: string = '';
    duration: string = '';
    country: string = '';
    description: string = '';
    overallRate: number = 0;
    imageUrl: File | undefined;
  }
  